Modify MaterialLevel property
Disable direct property output for cake materials (e.g. m@cake[level=0] will not be seen). Determining the bites a cake has left is not in the scope of the MaterialLevel property

Add additional meta to clarify that we have a dedicated property for cake bites

Allow both "max_" and "maximum_" prefixes

Disable property output for when the level is 0. The level being 0 is a trivial case